power: pmic: sandbox: support 256-register devices
Use unsigned integers for the register count and buffer size.
Eight-bit values wrap to zero for PMICs with 256 registers,
preventing the emulator from providing their register space.

+++ drivers/power/pmic/i2c_pmic_emul.c
@@ -21,9 +21,9 @@
  */
 struct sandbox_i2c_pmic_plat_data {
 	u8 rw_reg, rw_idx;
-	u8 reg_count;
+	uint reg_count;
 	u8 trans_len;
-	u8 buf_size;
+	uint buf_size;
 	u8 *reg;
 };
